Tiny typo fix
Hi there! I came across this post on your blog when I was looking for a way to set a timeout for some promises in a JS/TS project I'm working on. Your post was really helpful and allowed me do exactly what I was hoping for. On top of that, your explanation was detailed but concise which made understanding the solution you proposed very easy. I came across a small typo so I thought I'd make this PR in case you wanted it to be fixed.